Submodules failed, welcome nested git repos.
The sad truth about submodules is that you can NOT commit into them - only pull
and merge. Thus, the only sane way to do journalling is to create a directory
inside your repo, copy journal.config into .git and include it in .git/config.

You can't share (e.g. track history of) git configuration, so, for now, all that
we can do is to rely on manual labor.