Design and language decisions
When we're working with git-journal we want to know that we in journal subrepo.
Otherwise we might accidentaly commit to superrepo and that's bad.

We have to discover current repo and we can do this with libgit2. With this
library we can live like gods. Unfortunately, bash is so lame it doesn't
supported, so we'll use python. Yay!

I would like to use haskell or go for this, but don't want to deal with
distribution of binaries or having haskell dependencies (for runhaskell).